fix: move functions into separate project
Also reorganize other non-app code.

The v2 deploy was trying to do a vite build.
