+++ functions/common/README.md
@@ -1,0 +1,1 @@
+This folder needs to be within the `functions` folder because firebase deploy will only deploy the `functions` folder. This code is also used by the app.
