+++ functions/common/auth.js
