+++ functions/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "functions",
+  "private": true,
+  "description": "Cloud Functions for Firebase",
+  "type": "module",
+  "main": "index.js",
+  "scripts": {
+    "deploy": "firebase deploy --only functions",
+    "dev:firebase": "export GOOGLE_APPLICATION_CREDENTIALS=\"$PWD/functions-key-dev.json\" && firebase emulators:start --only functions",
+    "logs": "firebase functions:log",
+    "serve": "firebase emulators:start --only functions",
+    "shell": "firebase functions:shell",
+    "start": "npm run shell"
+  },
+  "dependencies": {
+    "@google-cloud/secret-manager": "^4.2.2",
+    "firebase-admin": "^11.8.0",
+    "firebase-functions": "^4.4.0",
+    "googleapis": "^118.0.0",
+    "yup": "^1.1.1"
+  },
+  "devDependencies": {
+    "firebase-functions-test": "^3.1.0"
+  },
+  "engines": {
+    "node": "18"
+  }
+}
