+++ package.json
@@ -6,18 +6,17 @@
   "license": "MIT",
   "author": "UGRC",
   "type": "module",
-  "main": "./src/functions/index.js",
   "scripts": {
     "build": "vite build",
     "build:dev": "npm run build:dev:update-remote-config-defaults && npm run build && npm run build:storybook",
-    "build:dev:update-remote-config-defaults": "export GOOGLE_APPLICATION_CREDENTIALS=\"$PWD/src/functions/functions-key-dev.json\" && node src/build-scripts/updateRemoteConfigDefaults.js",
+    "build:dev:update-remote-config-defaults": "export GOOGLE_APPLICATION_CREDENTIALS=\"$PWD/functions/functions-key-dev.json\" && node build-scripts/updateRemoteConfigDefaults.js",
     "build:prod": "npm run build",
     "build:stage": "npm run build && npm run build:storybook",
     "build:storybook": "storybook build --output-dir ./dist/storybook",
-    "build:update-remote-config-defaults": "node src/build-scripts/updateRemoteConfigDefaults.js",
+    "build:update-remote-config-defaults": "node build-scripts/updateRemoteConfigDefaults.js",
     "copy:arcgis": "cpy \"./node_modules/@arcgis/core/assets/**\" ./public/assets",
     "dev": "cross-env NODE_ENV=development run-p dev:*",
-    "dev:firebase": "export GOOGLE_APPLICATION_CREDENTIALS=\"$PWD/src/functions/functions-key-dev.json\" && firebase emulators:start --import .emulator-data --only functions",
+    "dev:firebase": "export GOOGLE_APPLICATION_CREDENTIALS=\"$PWD/functions/functions-key-dev.json\" && firebase emulators:start --import .emulator-data --only functions",
     "dev:vite": "vite",
     "format": "prettier --write .",
     "lint": "eslint ./ --ext .js,.jsx --ignore-path ./.eslintignore",
@@ -85,7 +84,6 @@
   },
   "dependencies": {
     "@arcgis/core": "^4.26.5",
-    "@google-cloud/secret-manager": "^4.2.2",
     "@radix-ui/react-accessible-icon": "^1.0.2",
     "@radix-ui/react-accordion": "^1.1.1",
     "@radix-ui/react-checkbox": "^1.0.3",
@@ -96,7 +94,6 @@
     "@xstate/react": "^3.2.2",
     "clsx": "^1.2.1",
     "firebase-admin": "^11.8.0",
-    "firebase-functions": "^4.4.0",
     "googleapis": "^118.0.0",
     "localforage": "^1.10.0",
     "react": "^18.2.0",
